Fix randomization to be more truly random
Earlier, we did the following:
1. Sort unlabeled keys
2. Shuffle unlabeled keys with a fixed seed.

The above process is problematic, as (2) seems to end up with a less
than random sequence. E.g., the element at index k often (maybe always)
seems to end up at the same index k', which means in practice we see the
keys in a somewhat-sorted order.